Rename README.md to README
 Interplay between dopamine D2 and insulin receptors in the central amygdala balances compulsive eating 
 Bokyeong Kim, Yoon Jeon, Ho Lee, Jung Hyun Pyo, Joung-Hun Kim, Je Kyung Seong, and Ja-Hyun Baik

Code file and demo dataset used in this study. 
You can use the code with the demo data to reproduce the expected figures.
It also has source data and instructions to use software to analyze fiber photometry data in the research.
Additional datasets or descriptions will be provided by the corresponding author upon reasonable request.